requirements.txt
beautifulsoup4==4.11.1
billboard.py==7.0.0
pytrends==4.7.1
requests==2.28.1
streamlit==1.17.0
